feat: 필터 구역에 슬라이드 기능 적용
슬라이드 버튼 클릭시 좌우 스크롤 + sticky 적용
